requirements.txt
kafka-python==2.0.2
requests==2.28.1
psycopg2-binary==2.9.5
python-dateutil==2.8.2
tornado>=6.0.0
